Correct ambiguous 'ls' command
In the "Copying and Moving Files" section the '-F' option is passed to
'ls' to show two files and two directories.  The next command moves one
of the files to a directory and the following example shows a bare 'ls'
but with one of the directories with a trailing '/' character (the
result of the '-F' option).

This patch removes the ambiguity by showing the result when '-F' is
passed to 'ls' by adding the option to the command and showing both
directories with trailing '/' characters.

This ambiguity is also present in the original book available from:
https://www.oreilly.com/openbook/utp/