Remove extraneous " lines" from Hb reference
The value of register Hb controls the heading levels that will have a
line break after the heading.  The addition of " lines" implies that
the value controls the number of blank lines which is not the case per
original AT&T MM documentation:

https://github.com/n-t-roff/DWB3.3/blob/562bb8615116f46077ca03e8561323cc7a89673e/doc/mm/mm.04#L257

Unfortunately, this erroneous text also appears in the original UTP and
confuses the intended use of the register.